[Hotfix] #12 - (1)번 반영
(1) 건강데이터 통계 보러 가기 삭제 => 기존 걸음수/수면/심장박독수/활동량 보여주는 박스를 위에를 조금 더 길게 하고, “통계 보기” 텍스트를 추가, 전체 박스를 클릭 가능하게 변경
